move scaffold/ resources into scaffold/resources
