+++ pkg/scaffold/resources/patch_test.go
@@ -1,4 +1,4 @@
-package scaffold
+package resources
 
 import (
 	"encoding/json"
